Refactored package-structure
