+++ src/main/java/fi/hsl/transitdata/tripupdate/application/IMessageProcessor.java
@@ -1,4 +1,4 @@
-package fi.hsl.transitdata.tripupdate;
+package fi.hsl.transitdata.tripupdate.application;
 
 import org.apache.pulsar.client.api.Message;
