+++ src/main/java/fi/hsl/transitdata/tripupdate/processing/DepartureProcessor.java
@@ -1,7 +1,8 @@
-package fi.hsl.transitdata.tripupdate;
+package fi.hsl.transitdata.tripupdate.processing;
 
 import com.google.protobuf.InvalidProtocolBufferException;
 import fi.hsl.common.transitdata.proto.PubtransTableProtos;
+import fi.hsl.transitdata.tripupdate.models.StopEvent;
 import org.apache.pulsar.client.api.Message;
 
 public class DepartureProcessor extends BaseProcessor {
