+++ src/test/java/fi/hsl/transitdata/tripupdate/gtfsrt/GtfsRtValidatorTest.java
@@ -1,6 +1,8 @@
-package fi.hsl.transitdata.tripupdate;
+package fi.hsl.transitdata.tripupdate.gtfsrt;
 
 import com.google.transit.realtime.GtfsRealtime;
+import fi.hsl.transitdata.tripupdate.MockDataFactory;
+import fi.hsl.transitdata.tripupdate.models.StopEvent;
 import org.junit.Test;
 
 import java.util.Arrays;
