+++ src/test/java/fi/hsl/transitdata/tripupdate/processing/TripUpdateProcessorTest.java
@@ -1,7 +1,9 @@
-package fi.hsl.transitdata.tripupdate;
+package fi.hsl.transitdata.tripupdate.processing;
 
 import com.google.transit.realtime.GtfsRealtime;
 import fi.hsl.common.transitdata.proto.PubtransTableProtos;
+import fi.hsl.transitdata.tripupdate.MockDataFactory;
+import fi.hsl.transitdata.tripupdate.models.StopEvent;
 import org.junit.Test;
 
 import java.util.List;
